+++ backend/server/index.ts
@@ -1,3 +1,14 @@
-import app from "./app";
+import { app, activate_db } from './app';
 
-app.listen(8080, () => console.log("Server is running on port 8080"));
+const startServer = async () => {
+  try {
+    await activate_db();
+
+    const PORT: number = 3000;
+    app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
+  } catch (error) {
+    console.error('Failed to connect to the database', error);
+  }
+};
+
+startServer();
